pyproject.toml
[project]
name = "test-automation-practice"
version = "0.1.0"
dependencies = [
    "pytest",
    "pytest-playwright",
    "Faker",
    "pytest-html",
    "pytest-metadata",
    "pytest-xdist"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--strict-markers"
